Move bluetooth logic to separate file
