+++ main/inc/bluetooth.h
@@ -1,4 +1,3 @@
-
 /*
  * Licensed to the Apache Software Foundation (ASF) under one
  * or more contributor license agreements.  See the NOTICE file
@@ -21,27 +20,21 @@
 #ifndef H_BLEPRPH_
 #define H_BLEPRPH_
 
-#include "syscfg/syscfg.h"
-#include "modlog/modlog.h"
-#include "nimble/ble.h"
-#include <stdbool.h>
 #ifdef __cplusplus
 extern "C" {
 #endif
 
+#include <stdbool.h>
+#include "nimble/ble.h"
+#include "modlog/modlog.h"
+
 struct ble_hs_cfg;
 struct ble_gatt_register_ctxt;
 
-/** GATT server. */
-#define GATT_SVR_SVC_ALERT_UUID 0x1811
-#define GATT_SVR_CHR_SUP_NEW_ALERT_CAT_UUID 0x2A47
-#define GATT_SVR_CHR_NEW_ALERT 0x2A46
-#define GATT_SVR_CHR_SUP_UNR_ALERT_CAT_UUID 0x2A48
-#define GATT_SVR_CHR_UNR_ALERT_STAT_UUID 0x2A45
-#define GATT_SVR_CHR_ALERT_NOT_CTRL_PT 0x2A44
-
-void gatt_svr_register_cb(struct ble_gatt_register_ctxt *ctxt, void *arg);
-int gatt_svr_init(void);
+/**
+ * Initializes and starts the Bluetooth task.
+ */
+void bluetooth_start(void);
 
 #ifdef __cplusplus
 }
