+++ main/src/main.c
@@ -8,341 +8,13 @@
 #include "esp_log.h"
 #include "freertos/FreeRTOS.h"
 #include "freertos/task.h"
-#include "host/ble_gap.h"
-#include "host/ble_hs.h"
-#include "host/util/util.h"
 #include "util.h"
 #include "scli.h"
 #include "mod_lvgl.h"
-#include "nimble/nimble_port.h"
-#include "nimble/nimble_port_freertos.h"
 #include "nvs_flash.h"
-#include "services/gap/ble_svc_gap.h"
 
 
 static const char *TAG = "MAIN";
-
-static void bleprph_advertise(void);
-
-static uint8_t own_addr_type; // Can be change into random address
-
-static void bleprph_on_sync(void) {
-
-  int rc;
-
-  rc = ble_hs_util_ensure_addr(0);
-  assert(rc == 0);
-
-  /* Figure out address to use while advertising (no privacy for now) */
-  rc = ble_hs_id_infer_auto(0, &own_addr_type);
-  if (rc != 0) {
-    MODLOG_DFLT(ERROR, "error determining address type; rc=%d\n", rc);
-    return;
-  }
-
-  /* Printing ADDR */
-  uint8_t addr_val[6] = {0};
-  rc = ble_hs_id_copy_addr(own_addr_type, addr_val, NULL);
-
-  MODLOG_DFLT(INFO, "Device Address: ");
-  print_addr(addr_val);
-  MODLOG_DFLT(INFO, "\n");
-  /* Begin advertising. */
-  bleprph_advertise();
-}
-
-static void bleprph_on_reset(int reason) {
-  MODLOG_DFLT(ERROR, "Resetting state; reason=%d\n", reason);
-}
-
-/**
- * Logs information about a connection to the console.
- */
-static void bleprph_print_conn_desc(struct ble_gap_conn_desc *desc) {
-  MODLOG_DFLT(INFO,
-              "handle=%d our_ota_addr_type=%d our_ota_addr=", desc->conn_handle,
-              desc->our_ota_addr.type);
-  print_addr(desc->our_ota_addr.val);
-  MODLOG_DFLT(INFO,
-              " our_id_addr_type=%d our_id_addr=", desc->our_id_addr.type);
-  print_addr(desc->our_id_addr.val);
-  MODLOG_DFLT(
-      INFO, " peer_ota_addr_type=%d peer_ota_addr=", desc->peer_ota_addr.type);
-  print_addr(desc->peer_ota_addr.val);
-  MODLOG_DFLT(INFO,
-              " peer_id_addr_type=%d peer_id_addr=", desc->peer_id_addr.type);
-  print_addr(desc->peer_id_addr.val);
-  MODLOG_DFLT(INFO,
-              " conn_itvl=%d conn_latency=%d supervision_timeout=%d "
-              "encrypted=%d authenticated=%d bonded=%d\n",
-              desc->conn_itvl, desc->conn_latency, desc->supervision_timeout,
-              desc->sec_state.encrypted, desc->sec_state.authenticated,
-              desc->sec_state.bonded);
-}
-
-/**
- * The nimble host executes this callback when a GAP event occurs.  The
- * application associates a GAP event callback with each connection that forms.
- * bleprph uses the same callback for all connections.
- *
- * @param event                 The type of event being signalled.
- * @param ctxt                  Various information pertaining to the event.
- * @param arg                   Application-specified argument; unused by
- *                                  bleprph.
- *
- * @return                      0 if the application successfully handled the
- *                                  event; nonzero on failure.  The semantics
- *                                  of the return code is specific to the
- *                                  particular GAP event being signalled.
- */
-static int bleprph_gap_event(struct ble_gap_event *event, void *arg) {
-  struct ble_gap_conn_desc desc;
-  int rc;
-
-  switch (event->type) {
-
-  case BLE_GAP_EVENT_CONNECT:
-    /* A new connection was established or a connection attempt failed. */
-    MODLOG_DFLT(INFO, "connection %s; status=%d ",
-                event->connect.status == 0 ? "established" : "failed",
-                event->connect.status);
-    if (event->connect.status == 0) {
-      rc = ble_gap_conn_find(event->connect.conn_handle, &desc);
-      assert(rc == 0);
-      bleprph_print_conn_desc(&desc);
-
-      /* Force Security Initiation to trigger Pairing on Phone */
-      rc = ble_gap_security_initiate(event->connect.conn_handle);
-      if (rc != 0) {
-          MODLOG_DFLT(INFO, "Security initiate failed; rc=%d\n", rc);
-      }
-    }
-    MODLOG_DFLT(INFO, "\n");
-
-    if (event->connect.status != 0) {
-      /* Connection failed; resume advertising. */
-      bleprph_advertise();
-    }
-
-    return 0;
-
-  case BLE_GAP_EVENT_DISCONNECT:
-    MODLOG_DFLT(INFO, "disconnect; reason=%d ", event->disconnect.reason);
-    bleprph_print_conn_desc(&event->disconnect.conn);
-    MODLOG_DFLT(INFO, "\n");
-
-    /* Connection terminated; resume advertising. */
-#if CONFIG_EXAMPLE_EXTENDED_ADV
-    ext_bleprph_advertise();
-#else
-    bleprph_advertise();
-#endif
-    return 0;
-
-  case BLE_GAP_EVENT_CONN_UPDATE:
-    /* The central has updated the connection parameters. */
-    MODLOG_DFLT(INFO, "connection updated; status=%d ",
-                event->conn_update.status);
-    rc = ble_gap_conn_find(event->conn_update.conn_handle, &desc);
-    assert(rc == 0);
-    bleprph_print_conn_desc(&desc);
-    MODLOG_DFLT(INFO, "\n");
-    return 0;
-
-  case BLE_GAP_EVENT_ADV_COMPLETE:
-    MODLOG_DFLT(INFO, "advertise complete; reason=%d",
-                event->adv_complete.reason);
-#if CONFIG_EXAMPLE_EXTENDED_ADV
-    ext_bleprph_advertise();
-#else
-    bleprph_advertise();
-#endif
-    return 0;
-
-  case BLE_GAP_EVENT_ENC_CHANGE:
-    /* Encryption has been enabled or disabled for this connection. */
-    MODLOG_DFLT(INFO, "encryption change event; status=%d ",
-                event->enc_change.status);
-    rc = ble_gap_conn_find(event->enc_change.conn_handle, &desc);
-    assert(rc == 0);
-    bleprph_print_conn_desc(&desc);
-    MODLOG_DFLT(INFO, "\n");
-    return 0;
-
-  case BLE_GAP_EVENT_NOTIFY_TX:
-    MODLOG_DFLT(INFO,
-                "notify_tx event; conn_handle=%d attr_handle=%d "
-                "status=%d is_indication=%d",
-                event->notify_tx.conn_handle, event->notify_tx.attr_handle,
-                event->notify_tx.status, event->notify_tx.indication);
-    return 0;
-
-  case BLE_GAP_EVENT_SUBSCRIBE:
-    MODLOG_DFLT(INFO,
-                "subscribe event; conn_handle=%d attr_handle=%d "
-                "reason=%d prevn=%d curn=%d previ=%d curi=%d\n",
-                event->subscribe.conn_handle, event->subscribe.attr_handle,
-                event->subscribe.reason, event->subscribe.prev_notify,
-                event->subscribe.cur_notify, event->subscribe.prev_indicate,
-                event->subscribe.cur_indicate);
-    return 0;
-
-  case BLE_GAP_EVENT_MTU:
-    MODLOG_DFLT(INFO, "mtu update event; conn_handle=%d cid=%d mtu=%d\n",
-                event->mtu.conn_handle, event->mtu.channel_id,
-                event->mtu.value);
-    return 0;
-
-  case BLE_GAP_EVENT_REPEAT_PAIRING:
-    /* We already have a bond with the peer, but it is attempting to
-     * establish a new secure link.  This app sacrifices security for
-     * convenience: just throw away the old bond and accept the new link.
-     */
-
-    /* Delete the old bond. */
-    rc = ble_gap_conn_find(event->repeat_pairing.conn_handle, &desc);
-    assert(rc == 0);
-    ble_store_util_delete_peer(&desc.peer_id_addr);
-
-    /* Return BLE_GAP_REPEAT_PAIRING_RETRY to indicate that the host should
-     * continue with the pairing operation.
-     */
-    return BLE_GAP_REPEAT_PAIRING_RETRY;
-
-  case BLE_GAP_EVENT_PASSKEY_ACTION:
-    ESP_LOGI(TAG, "PASSKEY_ACTION_EVENT started");
-    struct ble_sm_io pkey = {0};
-    int key = 0;
-
-    if (event->passkey.params.action == BLE_SM_IOACT_DISP) {
-      pkey.action = event->passkey.params.action;
-      pkey.passkey = 123456; // This is the passkey to be entered on peer
-      ESP_LOGI(TAG, "Enter passkey %" PRIu32 "on the peer side", pkey.passkey);
-      rc = ble_sm_inject_io(event->passkey.conn_handle, &pkey);
-      ESP_LOGI(TAG, "ble_sm_inject_io result: %d", rc);
-    } else if (event->passkey.params.action == BLE_SM_IOACT_NUMCMP) {
-      ESP_LOGI(TAG, "Passkey on device's display: %" PRIu32,
-               event->passkey.params.numcmp);
-      ESP_LOGI(TAG, "Accept or reject the passkey through console in this "
-                    "format -> key Y or key N");
-      pkey.action = event->passkey.params.action;
-      if (scli_receive_key(&key)) {
-        pkey.numcmp_accept = key;
-      } else {
-        pkey.numcmp_accept = 0;
-        ESP_LOGE(TAG, "Timeout! Rejecting the key");
-      }
-      rc = ble_sm_inject_io(event->passkey.conn_handle, &pkey);
-      ESP_LOGI(TAG, "ble_sm_inject_io result: %d", rc);
-    } else if (event->passkey.params.action == BLE_SM_IOACT_OOB) {
-      static uint8_t tem_oob[16] = {0};
-      pkey.action = event->passkey.params.action;
-      for (int i = 0; i < 16; i++) {
-        pkey.oob[i] = tem_oob[i];
-      }
-      rc = ble_sm_inject_io(event->passkey.conn_handle, &pkey);
-      ESP_LOGI(TAG, "ble_sm_inject_io result: %d", rc);
-    } else if (event->passkey.params.action == BLE_SM_IOACT_INPUT) {
-      ESP_LOGI(TAG,
-               "Enter the passkey through console in this format-> key 123456");
-      pkey.action = event->passkey.params.action;
-      if (scli_receive_key(&key)) {
-        pkey.passkey = key;
-      } else {
-        pkey.passkey = 0;
-        ESP_LOGE(TAG, "Timeout! Passing 0 as the key");
-      }
-      rc = ble_sm_inject_io(event->passkey.conn_handle, &pkey);
-      ESP_LOGI(TAG, "ble_sm_inject_io result: %d", rc);
-    }
-    return 0;
-
-  case BLE_GAP_EVENT_AUTHORIZE:
-    MODLOG_DFLT(INFO,
-                "authorize event: conn_handle=%d attr_handle=%d is_read=%d",
-                event->authorize.conn_handle, event->authorize.attr_handle,
-                event->authorize.is_read);
-
-    /* The default behaviour for the event is to reject authorize request */
-    event->authorize.out_response = BLE_GAP_AUTHORIZE_REJECT;
-    return 0;
-  }
-  return 0;
-}
-
-/**
- * Enables advertising with the following parameters:
- *     o General discoverable mode.
- *     o Undirected connectable mode.
- */
-static void bleprph_advertise(void) {
-  struct ble_gap_adv_params adv_params;
-  struct ble_hs_adv_fields fields;
-  const char *name;
-  int rc;
-
-  /**
-   *  Set the advertisement data included in our advertisements:
-   *     o Flags (indicates advertisement type and other general info).
-   *     o Advertising tx power.
-   *     o Device name.
-   *     o 16-bit service UUIDs (alert notifications).
-   */
-
-  memset(&fields, 0, sizeof fields);
-
-  /* Advertise two flags:
-   *     o Discoverability in forthcoming advertisement (general)
-   *     o BLE-only (BR/EDR unsupported).
-   */
-  fields.flags = BLE_HS_ADV_F_DISC_GEN | BLE_HS_ADV_F_BREDR_UNSUP;
-
-  /* Indicate that the TX power level field should be included; have the
-   * stack fill this value automatically.  This is done by assigning the
-   * special value BLE_HS_ADV_TX_PWR_LVL_AUTO.
-   */
-  fields.tx_pwr_lvl_is_present = 1;
-  fields.tx_pwr_lvl = BLE_HS_ADV_TX_PWR_LVL_AUTO;
-
-  name = ble_svc_gap_device_name();
-  fields.name = (uint8_t *)name;
-  fields.name_len = strlen(name);
-  fields.name_is_complete = 1;
-
-  static const ble_uuid16_t adv_uuids[] = {
-      BLE_UUID16_INIT(GATT_SVR_SVC_ALERT_UUID)};
-  fields.uuids16 = adv_uuids;
-  fields.num_uuids16 = 1;
-  fields.uuids16_is_complete = 1;
-
-  rc = ble_gap_adv_set_fields(&fields);
-  if (rc != 0) {
-    MODLOG_DFLT(ERROR, "error setting advertisement data; rc=%d\n", rc);
-    return;
-  }
-
-  /* Begin advertising. */
-  memset(&adv_params, 0, sizeof adv_params);
-  adv_params.conn_mode = BLE_GAP_CONN_MODE_UND;
-  adv_params.disc_mode = BLE_GAP_DISC_MODE_GEN;
-  rc = ble_gap_adv_start(own_addr_type, NULL, BLE_HS_FOREVER, &adv_params,
-                         bleprph_gap_event, NULL);
-  if (rc != 0) {
-    MODLOG_DFLT(ERROR, "error enabling advertisement; rc=%d\n", rc);
-    return;
-  }
-}
-
-void ble_store_config_init(void); // What's the purpose of this function need to
-                                  // check the walk through again
-
-void bleprph_host_task(void *param) {
-  ESP_LOGI(TAG, "BLE Host Task Started");
-  /* This function will return only when nimble_port_stop() is executed */
-  nimble_port_run();
-
-  nimble_port_freertos_deinit();
-}
 
 /* -------------------- app_main -------------------- */
 
@@ -359,38 +31,7 @@ void app_main(void) {
   }
   ESP_ERROR_CHECK(ret);
 
-  ret = nimble_port_init();
-  if (ret != ESP_OK) {
-    ESP_LOGE(TAG, "Failed to init nimble %d ", ret);
-    return;
-  }
-
-  /* Initialize the NimBLE host configuration. */
-  ble_hs_cfg.reset_cb = bleprph_on_reset;
-  ble_hs_cfg.sync_cb = bleprph_on_sync;
-  ble_hs_cfg.gatts_register_cb = gatt_svr_register_cb;
-  ble_hs_cfg.store_status_cb = ble_store_util_status_rr;
-
-  ble_hs_cfg.sm_io_cap = CONFIG_EXAMPLE_IO_TYPE;
-  ble_hs_cfg.sm_sc = 1;     // Enable Secure Connections
-  ble_hs_cfg.sm_mitm = 1;   // Force Man-In-The-Middle protection (triggers Passkey)
-  ble_hs_cfg.sm_bonding = 1; // Enable Bonding
-
-  /* Stores the IRK */
-  /* ble_hs_cfg.sm_our_key_dist |= BLE_SM_PAIR_KEY_DIST_ID; */
-  /* ble_hs_cfg.sm_their_key_dist |= BLE_SM_PAIR_KEY_DIST_ID; */
-
-  rc = gatt_svr_init();
-  assert(rc == 0);
-
-  /* Set the default device name. */
-  rc = ble_svc_gap_device_name_set("nimble-test");
-  assert(rc == 0);
-
-  /* XXX Need to have template for store */
-  ble_store_config_init();
-
-  nimble_port_freertos_init(bleprph_host_task);
+  bluetooth_start();
 
   /* Initialize command line interface to accept input from user */
   rc = scli_init();
